requirements.txt
python-dotenv==1.0.0
crewai==0.11.0
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
requests==2.31.0
chardet==5.2.0
certifi==2024.2.2 